Update db.js to check spend limit with on new visit
Call `checkMaxAndEmail` from `addVisit` so every time the total amount
spent on a provider is augmented, the total amount spent is checked and
an email is sent to the rep and the boss if they have gone over the spend
limit. The 'from' email must be [verified](https://sendgrid.com/docs/for-developers/sending-email/sender-identity/) before emails
can be sent.